[00003-0] Implementing PCA9685 support
The I2C driver (i2c.h, i2c.c files) was modified in order to work with
diffirent I2C devices. Added new argument I2C_TypeDef* I2Cx.

The PCA9685 (later as PCA) driver is supposed to be used with I2C1 device, which is
denoted in PCA9685_I2C variable definition. To init the PCA9685 device
a user have to call the pca9685_init function and pass to it required
frequency for the PCA. The user can set the PWM power for a separete channel
with a pca9685_set_pwm function.